third_party/WebKit/LayoutTests/fast/js/primitive-property-access-edge-cases-expected.txt

This page tests for assertion failures in edge cases of property lookup on primitive values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS checkGet(1, Number) is true
PASS checkGet('hello', String) is true
PASS checkGet(true, Boolean) is true
FAIL checkSet(1, Number) should be true. Was false.
FAIL checkSet('hello', String) should be true. Was false.
FAIL checkSet(true, Boolean) should be true. Was false.
PASS checkGetStrict(1, Number) is true
PASS checkGetStrict('hello', String) is true
PASS checkGetStrict(true, Boolean) is true
FAIL checkSetStrict(1, Number) should be true. Was false.
FAIL checkSetStrict('hello', String) should be true. Was false.
FAIL checkSetStrict(true, Boolean) should be true. Was false.
PASS checkRead(1, Number) is true
PASS checkRead('hello', String) is true
PASS checkRead(true, Boolean) is true
PASS checkWrite(1, Number) is true
PASS checkWrite('hello', String) is true
PASS checkWrite(true, Boolean) is true
PASS checkReadStrict(1, Number) is true
PASS checkReadStrict('hello', String) is true
PASS checkReadStrict(true, Boolean) is true
FAIL checkWriteStrict(1, Number) should throw an exception. Was true.
FAIL checkWriteStrict('hello', String) should throw an exception. Was true.
FAIL checkWriteStrict(true, Boolean) should throw an exception. Was true.
PASS checkNumericGet(1, Number) is true
PASS checkNumericGet('hello', String) is true
PASS checkNumericGet(true, Boolean) is true
FAIL checkNumericSet(1, Number) should be true. Was false.
FAIL checkNumericSet('hello', String) should be true. Was false.
FAIL checkNumericSet(true, Boolean) should be true. Was false.
FAIL checkNumericGetStrict(1, Number) should be true. Was false.
FAIL checkNumericGetStrict('hello', String) should be true. Was false.
FAIL checkNumericGetStrict(true, Boolean) should be true. Was false.
FAIL checkNumericSetStrict(1, Number) should be true. Was false.
FAIL checkNumericSetStrict('hello', String) should be true. Was false.
FAIL checkNumericSetStrict(true, Boolean) should be true. Was false.
PASS checkNumericRead(1, Number) is true
PASS checkNumericRead('hello', String) is true
PASS checkNumericRead(true, Boolean) is true
PASS checkNumericWrite(1, Number) is true
PASS checkNumericWrite('hello', String) is true
PASS checkNumericWrite(true, Boolean) is true
PASS checkNumericReadStrict(1, Number) is true
PASS checkNumericReadStrict('hello', String) is true
PASS checkNumericReadStrict(true, Boolean) is true
FAIL checkNumericWriteStrict(1, Number) should throw an exception. Was true.
FAIL checkNumericWriteStrict('hello', String) should throw an exception. Was true.
FAIL checkNumericWriteStrict(true, Boolean) should throw an exception. Was true.
PASS didNotCrash is true
PASS successfullyParsed is true

TEST COMPLETE
